Add files via upload
This project contains the process for creating a model to more accurately set a spread for NFL games. 

This model accurately predicted the winner for nearly 80% of all NFL games, in addition to setting more accurate spreads in week 13 than sports books at over an 80% rate. The model contained is an ensemble of several different machine learning models that includes not only publicly available statistics for NFL teams, but also a sentiment analysis using user posts on the 'r/NFL' subreddit.